Common: fixed MemoryStream::ReadByte()
ReadByte is supposed to return *unsigned char* cast to int32, was *signed char* instead which could cause erroneous end of stream result.

+++ Common/util/memorystream.cpp
@@ -115,7 +115,7 @@ size_t MemoryStream::Read(void *buffer, size_t size)
 int32_t MemoryStream::ReadByte()
 {
     if (EOS()) { return -1; }
-    return _cbuf[(size_t)(_pos++)];
+    return static_cast<uint8_t>(_cbuf[(size_t)(_pos++)]);
 }
 
 size_t MemoryStream::Write(const void *buffer, size_t size)
